Refactor message-listener functionality
