+++ native/src/main/java/io/ballerina/stdlib/java.jms/listener/ListenerCallback.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.java.jms;
+package io.ballerina.stdlib.java.jms.listener;
 
 import io.ballerina.runtime.api.async.Callback;
 import io.ballerina.runtime.api.values.BError;
@@ -24,7 +24,7 @@ import io.ballerina.runtime.api.values.BError;
 /**
  * Callback code to be executed when the message-listener complete a message producing cycle to the ballerina service.
  */
-public class ConsumerCallback implements Callback {
+public class ListenerCallback implements Callback {
     @Override
     public void notifySuccess(Object o) {
         if (o instanceof BError) {
@@ -37,4 +37,5 @@ public class ConsumerCallback implements Callback {
         bError.printStackTrace();
         System.exit(1);
     }
+
 }
