+++ spotbugs-exclude.xml
@@ -17,7 +17,7 @@
   -->
 <FindBugsFilter>
     <Match>
-        <Class name="io.ballerina.stdlib.java.jms.ConsumerCallback"/>
+        <Class name="io.ballerina.stdlib.java.jms.listener.ListenerCallback"/>
         <Bug pattern="DM_EXIT"/>
     </Match>
 </FindBugsFilter>
